Move PreferenceTest to unit tests
